Style for V1.1.0 (#4)
* Fixed index and added footer

- Fixed grammar errors in index.html
- Added a footer in index.html and style_index.css

* Updated index HTML and CSS

Redesigned the index.html page.

* Updated layout for index and fixed translations

Added separators for index.html (needed for a future revision) and translated Italian texts to English. Fixed issue with footer going outside of the viewport. Added footer to admin.html.

* Separeted changelog and Readme

We have separeted the changelog out of readme for better reading into the index

* Footer changed

Now the footer will open a pop-up box for the visualization of README.md and LICENSE, now both files can be viewed directly within the browser.

* Added navbar to header

Added a horizontal navbar to the header, made a vertical navbar when the screen is smaller than 600px and fixed the footer hiding content.

* Fixed TOP/BOT cell

Fixed cell TOP/BOT having a higher colspan than normal and fixed the padding of the BOT button.

* Started Copy

Started developing of copy function for overlay links

* Fixed Modal

Fixed a bug for mobile devices, now you can close the popup of README.md

* Added copy function

Added copy function for overlay links via a button to the right of the overlay links.

Co-authored-by: TheTecnoKing <kristian.16122005@gmail.com>