Partial page added (parz.html)
the page has been added to allow the display of partials, this page contains a scoreboard divided into partials from 1 to max_inning (defined in the settings.json file the HTML page will be added later).
In the finished partial or in the partial with scores the score will be displayed underlined by a different background.
In this commit I add:
* parz.html the scoreboard page
* settings.json the settings for the scoreboard
* style_parz.css the style file for the display board