docs: 📝 Rewrite README
Completely rewrote the README, now with a new layout and new instructions. Also available in Italian. But it isn't complete, yet. There aren't any docs, the "Usage" section has to be written, and there are no images/screenshots to illustrate the software. Also, (optional) it would be nice to have a banner and a few badges, just to make it look a bit cooler.